WIP | Feat: Implement Emisor Receiver functionality

+++ src/main/kotlin/tech/noir/billsapp/bills/domain/model/Bill.kt
@@ -1,16 +1,16 @@
 package tech.noir.billsapp.bills.domain.model
 
 data class Bill(
-    val id: Int,
+    val id: Long,
     val uuid: String,
-    val billNumber : String,
-    val expeditionDate : String,
+    val billNumber: String,
+    val expeditionDate: String,
     val emisorUuid: String,
     val receiverUuid: String,
-    val concept : String,
-    val base : Double,
-    val iva : Double,
-    val total : Double,
+    val concept: String,
+    val base: Double,
+    val iva: Double,
+    val total: Double,
     val createdAt: String,
     val updatedAt: String,
     val deletedAt: String,
